+++ src/Symfony/Component/Console/CHANGELOG.md
@@ -14,6 +14,7 @@ CHANGELOG
  * Add support for `LockableTrait` in invokable commands
  * Deprecate returning a non-integer value from a `\Closure` function set via `Command::setCode()`
  * Mark `#[AsCommand]` attribute as `@final`
+ * Add `--no-progress` option to the commands
 
 7.2
 ---